add lexicon to deploy and start moving
