+++ packages/deploy/package.json
@@ -26,6 +26,7 @@
   "license": "ISC",
   "devDependencies": {
     "@inquirer/testing": "^2.1.53",
+    "@openfn/lexicon": "workspace:^",
     "@types/json-diff": "^1.0.3",
     "@types/jsonpath": "^0.2.4",
     "@types/mock-fs": "^4.13.4",
