+++ packages/deploy/src/types.ts
@@ -1,3 +1,5 @@
+import { Provisioner } from '@openfn/lexicon/lightning';
+
 export type StateJob = {
   id: string;
   name: string;
@@ -38,34 +40,9 @@ export type SpecKafkaConfiguration = {
   connect_timeout: number;
 };
 
-export type WebhookResponse = {
-  error_code: number | null;
-  success_code: number | null;
-};
+export type SpecTrigger = Omit<Provisioner.Trigger, 'id'>;
 
-export type WebhookReply = 'before_start' | 'after_completion';
-
-export type SpecTrigger = {
-  type: string;
-  cron_expression?: string;
-  cron_cursor_job?: string;
-  webhook_reply?: WebhookReply;
-  webhook_response?: WebhookResponse | null;
-  enabled?: boolean;
-  kafka_configuration?: SpecKafkaConfiguration;
-};
-
-export type StateTrigger = {
-  id: string;
-  type: string;
-  cron_expression?: string;
-  cron_cursor_job_id?: string | null;
-  webhook_reply?: WebhookReply;
-  webhook_response?: WebhookResponse | null;
-  delete?: boolean;
-  enabled?: boolean;
-  kafka_configuration?: StateKafkaConfiguration;
-};
+export type StateTrigger = Provisioner.Trigger;
 
 export type StateEdge = {
   id: string;
